requirements.txt
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.1
channels
channels-redis==3.4.1
charset-normalizer==2.0.12
class-registry==2.1.2
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
dateutils==0.6.12
Django==3.2.16
django-cors-headers==3.12.0
django-filter==22.1
django-jsonfield==1.4.1
django-restframework==0.0.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
filters==1.3.2
filters-django==1.0.5
gunicorn==20.1.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
msgpack==1.0.5
Pillow==9.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==23.1.1
pyotp==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.4.2
regex==2022.10.31
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.9
zope.interface==6.0
